# Refactor: Avoid redundant API calls