… ansible_collections/arista/cvp/plugins/module_utils/device_tools.py