## Change Summary

<!-- Enter short PR description -->
Improve performance of especially `cv_device_v3` by avoiding redundant API calls.

## Component(s) name

- `configlet_tools.py`
- `container_tools.py`
- `device_tools.py`

## Proposed changes
<!--- Describe your changes in detail -->
<!--- Describe data model implemented for new features -->

- Refactoring code to avoid duplicate API calls
- Implement `lru_cache` on some methods with API calls, to cache the result for repeated calls.
- Invalidate `lru_cache` for containers when containers are created/deleted.

## How to test
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->

Pass all CI.

Testing with small 6 device inventory in 3 containers. No imagebundles.

### Before
- Initial run - creating containers, adding configlets, moving devices (Creating tasks)
  **100 seconds**
- Subsequent run with no changes
  **60 seconds**

### After
- Initial run - creating containers, adding configlets, moving devices (Creating tasks)
  **66 seconds**
- Subsequent run with no changes
  **30 seconds**

## Checklist

### User Checklist

<!-- Add your own checklist using MD syntax and by replacing N/A -->
- N/A

### Repository Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code has been rebased from devel before I start
- [x] I have read the [**CONTRIBUTING**](https://avd.sh/en/latest/docs/contribution/overview.html) document.
- [x] My change requires a change to the documentation and documentation have been updated accordingly. (check the box if not applicable)
- [x] I have updated [molecule CI](https://github.com/aristanetworks/ansible-avd/tree/devel/ansible_collections/arista/avd/molecule) testing accordingly. (check the box if not applicable)