@rwedmonds 2023-03-14
# PR Test Results

| PR # | Test | Command | Output | Time |
| ---- | ---- | ------- | ------ | ---- |
| 592 | Initial run | ansible-playbook cv_device_v3/pr_592.yaml -i inventory.yaml | LAY [cv_container_v3] ****************</br></br>TASK [Configure containers on CloudVision] ****************</br>changed: [CloudVision]</br></br>TASK [Push configlets] ****************</br>changed: [CloudVision]</br></br>TASK [Move devices on CloudVision] ****************</br>changed: [CloudVision]</br></br>PLAY RECAP ****************</br>CloudVision                : ok=3    changed=3    unreachable=0    failed=0    skipped=0    rescued=0    ignored=0 | 68 seconds |
| | Subsequent run | Initial run | ansible-playbook cv_device_v3/pr_592.yaml -i inventory.yaml | PLAY [cv_container_v3] ****************</br></br>TASK [Configure containers on CloudVision] ****************</br>ok: [CloudVision]</br></br>TASK [Push configlets] ****************</br>changed: [CloudVision]</br></br>TASK [Move devices on CloudVision] ****************</br>ok: [CloudVision]</br></br>PLAY RECAP ****************</br>CloudVision                : ok=3    changed=1    unreachable=0    failed=0    skipped=0    rescued=0    ignored=0 | 19 seconds |